Move StreamLifecycleObserver to LLC
