--- stream-chat-android/src/main/java/com/getstream/sdk/chat/LifecycleHandler.java
@@ -1,7 +1,0 @@
-package com.getstream.sdk.chat;
-
-public interface LifecycleHandler {
-    void resume();
-
-    void stopped();
-}
